Fix user resource unit test are backward
Also add the unit test for `UserResolver.findByEmail`
`UserService.createPasswordResetLink`
