+++ src/resource/user/test/user.service.spec.ts
@@ -1,41 +1,72 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { UserResolver } from '../user.resolver';
+import { PrismaService } from '@common';
 import { UserService } from '../user.service';
 import mockUserData from './mockUserData';
+import * as crypto from 'crypto';
 
-describe('UserResolver', () => {
-  let resolver: UserResolver;
+describe('UserService', () => {
+  let userService: UserService;
 
-  const mockUserService = {
-    findAll: jest.fn(() => mockUserData),
-    findByUsername: jest.fn((_username: string) =>
-      mockUserData.filter(({ username }) => username === _username),
-    ),
+  const mockPrismaClient = {
+    user: {
+      create: jest.fn(({ data }) =>
+        Promise.resolve({ id: Math.random(), ...data }),
+      ),
+      findMany: jest.fn(() => mockUserData),
+      findFirst: jest.fn(({ where: { username: _username } }) =>
+        mockUserData.filter(({ username }) => username === _username),
+      ),
+      update: jest.fn(),
+    },
   };
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      providers: [UserResolver, UserService],
+      providers: [UserService, PrismaService],
     })
-      .overrideProvider(UserService)
-      .useValue(mockUserService)
+      .overrideProvider(PrismaService)
+      .useValue(mockPrismaClient)
       .compile();
 
-    resolver = module.get<UserResolver>(UserResolver);
+    userService = module.get<UserService>(UserService);
   });
 
   it('should be defined', () => {
-    expect(resolver).toBeDefined();
+    expect(userService).toBeDefined();
   });
 
-  it('should find all user', async () => {
-    expect(resolver.findAll()).toEqual(mockUserData);
+  it('should create user', async () => {
+    const newUser = mockUserData[0];
+    expect(await userService.create(newUser)).toEqual({
+      id: expect.any(Number),
+      ...newUser,
+    });
   });
 
-  it('should find user by username', async () => {
-    const user = mockUserData[1];
-    const expectedResponse = [user];
-    const username = user.username;
-    expect(resolver.findByUsername({ username })).toEqual(expectedResponse);
+  it('should find all users', async () => {
+    expect(await userService.findAll()).toBe(mockUserData);
+  });
+
+  it('should find users by username', async () => {
+    const user = mockUserData[0];
+    const input = user.username;
+    const output = [user];
+
+    expect(await userService.findByUsername(input)).toEqual(output);
+  });
+
+  it('should create password reset link', async () => {
+    const user = mockUserData[0];
+    // @Note: We can't really test if the resetToken was actually generated, so we are just mocking this function so that it don't fail the test
+    jest.spyOn(crypto, 'createHmac').mockImplementationOnce(() => {
+      return {
+        update: jest.fn().mockReturnThis(),
+        digest: jest.fn(),
+      } as unknown as crypto.Hmac;
+    });
+
+    expect(await userService.createPasswordResetLink(user.id)).toContain(
+      `${process.env.CLIENT_BASE_URL}/reset-password/`,
+    );
   });
 });
